обработка ошибок в main, перенес логику timer в view
