+++ task_2_3_1/src/main/java/vylegzhanin/snake/view/Main.java
@@ -5,6 +5,8 @@ import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.Alert.AlertType;
 import javafx.stage.Stage;
 
 /**
@@ -14,12 +16,21 @@ public class Main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        Parent root = FXMLLoader.load(
-            Objects.requireNonNull(getClass().getResource("/vylegzhanin/snake/MainView.fxml")));
-        primaryStage.setTitle("Snake Game OOP - Levels & Items");
-        primaryStage.setScene(new Scene(root));
-        primaryStage.setResizable(false);
-        primaryStage.show();
+        try {
+            Parent root = FXMLLoader.load(
+                Objects.requireNonNull(getClass().getResource("/vylegzhanin/snake/MainView.fxml")));
+            primaryStage.setTitle("Snake Game OOP - Levels & Items");
+            primaryStage.setScene(new Scene(root));
+            primaryStage.setResizable(false);
+            primaryStage.show();
+        } catch (Exception e) {
+            e.printStackTrace();
+            Alert alert = new Alert(AlertType.ERROR);
+            alert.setTitle("Error launching Game");
+            alert.setHeaderText("Could not load MainView.fxml");
+            alert.setContentText(e.getMessage());
+            alert.showAndWait();
+        }
     }
 
     public static void main(String[] args) {
